add file verification
Took 13 minutes

+++ write.go
@@ -1,11 +1,19 @@
 package go_filer
 
 import (
+	"fmt"
 	"io/ioutil"
 )
 
 func WriteFile(message string, filePath string) bool{
-	messageToByte := []byte (message)
-	_ = ioutil.WriteFile(filePath, messageToByte, 0644)
+	fileExists := FileExists(filePath)
+	if fileExists {
+		fmt.Println("The file already exists!")
+		return false
+	}else {
+		messageToByte := []byte (message)
+		_ = ioutil.WriteFile(filePath, messageToByte, 0644)
+
+	}
 	return true
 }
